Polished CSS and did phase 2 testing

+++ app.py
@@ -1,8 +1,6 @@
-from s import key, c_id, a_id
+from s import c_id, a_id
 from flask import Flask, render_template, redirect, flash, session, request, jsonify
-# request
 from flask_debugtoolbar import DebugToolbarExtension
-# from requests import session
 import requests
 req_session = requests.session
 from models import db, connect_db, Category_Game, Category, Game, User, Comment
@@ -21,7 +19,6 @@ debug = DebugToolbarExtension(app)
 connect_db(app)
 data=[]
 
-
 @app.route('/')
 def index():
     """Populates the login/sign-up page"""
@@ -29,6 +26,7 @@ def index():
 
 @app.route('/game_result', methods=["GET", "POST"])
 def game_result():
+    """Renders game resutls using the IGDB API, this IPO happens in the Search_logic def()"""
     games = Game.query.filter_by(favorite=True, user_id=session['user_id']).all()
     game = {"lookup_id": [g.api_id for g in games],
     "user_id": [g.user_id for g in games]
@@ -40,9 +38,9 @@ def game_result():
         return redirect('/game_result')
     return render_template('game_result.html', info=data, size=size, form=form, game=game)
 
-
 @app.route('/game-home',methods=["GET","POST"])
 def game_home():
+    """Renders the post login screen. Checks for saved games and created catergories. Also handels form subs"""
     games = Game.query.all()
     cat_form = AddCategoryForm()
     if "user_id" not in session:
@@ -53,7 +51,6 @@ def game_home():
     if form.validate_on_submit():
         data.append(search_logic(form,data,c_id,a_id))
         return redirect('/game_result')
-    
     elif cat_form.validate_on_submit():
         name = cat_form.name.data
         description = cat_form.description.data
@@ -72,6 +69,7 @@ def game_home():
 
 @app.route('/register', methods=['GET', 'POST'])
 def register_user():
+    """Creates new user and handels validation"""
     form = UserForm()
     if form.validate_on_submit():
         username = form.username.data
@@ -90,6 +88,8 @@ def register_user():
 
 @app.route('/game_info/<int:id>', methods=['GET', 'POST'])
 def game_info(id):
+    """Renders the selected game page providing Twitch live TV, youtube trailer/gameplay and live tweets from twitch.
+     uses the Twitch, IGDB, and giphy APIs"""
     if "user_id" not in session:
         flash("Please login!")
         return redirect('/')
@@ -100,24 +100,31 @@ def game_info(id):
         return redirect('/game-home')
     api_id = game.api_id
     info = search_by_id_logic(api_id,data,c_id,a_id)
-    img = info[0]["screenshots"]
-    video = info[0]["videos"][0]
-    vid_data = search_for_vid(video,data, c_id, a_id)
-    live_video = get_live_video(game.title)
-    print('57483975708932758930275983479')
-    print(live_video)
+    try:
+        img = info[0]["screenshots"]
+        video = info[0]["videos"][0]
+        vid_data = search_for_vid(video,data, c_id, a_id)
+        live_video = get_live_video(game.title, c_id, a_id)
+    except:
+        img = []
+        video = info[0]["videos"][0]
+        vid_data = search_for_vid(video,data, c_id, a_id)
+        live_video = get_live_video(game.title, c_id, a_id)
     return render_template('game_info.html', game=game, info=info, img=img, vid_data=vid_data, comment=comment, live_video=live_video)
 
 @app.route('/game_info/comment/<int:id>', methods=['POST'])
 def submit_comment(id):
+    """Lets user add notes for a specific game page that is saved in the DB"""
     comment = request.form['comments']
     new_comment = Comment(comment=comment, user_id=session['user_id'], game_id=id)
     db.session.add(new_comment)
     db.session.commit()
+    flash("Note Added")
     return redirect(f'/game_info/{id}')
 
 @app.route('/login', methods=['GET','POST'])
 def login_user():
+    """Lets exisiting users login and checks for valid user/name. Validation done with Bcrypt"""
     form = UserForm()
     if form.validate_on_submit():
         username = form.username.data
@@ -133,33 +140,31 @@ def login_user():
 
 @app.route('/logout', methods=["POST"])
 def logout_user():
+    """Removes user id from seesion to log out"""
     session.pop('user_id')
     flash("Logout successful!")
     return redirect ('/')
 
-
 @app.route('/get_fav', methods=["POST"])
 def get_fav():
+    """Listens for request incoming from JS axios and apends data to user home-page"""
     data = {
         "fav_input":"",
         "fav_name":"",
         "api_Id":"",
     }
-
-    
     if request.json['fav_input'] == 'on':
         title = request.json['fav_name']
         api_id = request.json['api_Id']
-        new_game = Game( title=title, genre="genre", company="company", summary='summary', url="url", favorite=True, user_id=session['user_id'], api_id=api_id )
+        new_game = Game( title=title, favorite=True, user_id=session['user_id'], api_id=api_id )
         session['fav_game'] = new_game.title
         db.session.add(new_game)
         db.session.commit()
-        print('good to go')
-        
     return jsonify(data) 
 
 @app.route('/delete_fav', methods=["POST"])
 def delete_fav():
+    """Listens for a delete request and removes games. Incoming request from JS Axios"""
     data = {
         "fav_input":"",
         "fav_name":""
@@ -167,46 +172,67 @@ def delete_fav():
     title = request.json['fav_name']
     game = Game.query.filter_by(title=title, user_id=session['user_id']).first()
     category_game = Category_Game.query.filter(Category_Game.game_id==game.id).first()
-    db.session.delete(category_game)
-    db.session.commit()
+    if category_game:
+        db.session.delete(category_game)
+        db.session.commit()
     all_comments = Comment.query.filter(Comment.game_id==game.id).all()
-    
     for a in all_comments:
         db.session.delete(a)
         db.session.commit()
     db.session.delete(game)
     db.session.commit()
     return jsonify(data)
-
 @app.route('/delete/<int:id>', methods=['POST'])
 def delete_from_home(id):
+    """Remove a game from all locations"""
     game = Game.query.get(id)
     all_comments = Comment.query.filter(Comment.game_id==game.id).all()
     for a in all_comments:
         db.session.delete(a)
         db.session.commit()
-    all_game = Category_Game.query.all()
-    category_game = Category_Game.query.filter(Category_Game.game_id==game.id).first()
-    db.session.delete(category_game)
-    db.session.commit()
+    category_game = Category_Game.query.filter(Category_Game.game_id==game.id).all()
+    for a in category_game:
+        if a.game_id == id:
+            db.session.delete(a)
+            db.session.commit()
     db.session.delete(game)
     db.session.commit()
+    flash('Game Removed Everywhere')
     return redirect('/game-home')
 
 @app.route('/delete/category/<int:id>', methods=['POST'])
 def delete_cat_from_home(id):
+    """Remove a custom category"""
     category = Category.query.get(id)
-    all_cat = Category_Game.query.all()
+    all_cat = Category_Game.query.filter(Category_Game.category_id==id).all()
     for a in all_cat:
         if a.category_id:
             db.session.delete(a)
             db.session.commit()
     db.session.delete(category)
     db.session.commit()
+    flash("Category Removed")
     return redirect('/game-home')
+
+@app.route('/delete/category/<int:id>/<int:game_id>')
+def delete_game_from_cat(id, game_id):
+    """Remove only a single game from one category"""
+    category_game = Category_Game.query.filter(Category_Game.category_id==id,Category_Game.game_id==game_id).first()
+    db.session.delete(category_game)
+    db.session.commit()
+    return redirect(f'/category/{id}')
+@app.route('/delete/comment/<int:id>/game/<int:game_id>', methods=['POST'])
+def delete_comment(id,game_id):
+    """Removes a note from a game page"""
+    comment = Comment.query.get(id)
+    db.session.delete(comment)
+    db.session.commit()
+    flash('Note Deleted')
+    return redirect(f'/game_info/{game_id}')
 
 @app.route('/category/game/<int:id>', methods=["POST"])
 def custom_cat(id):
+    """Allows user to make custom categories"""
     category_id = request.form['addCat']
     try:
         q = Category_Game.query.filter(Category_Game.category_id==category_id, Category_Game.game_id==id)
@@ -217,6 +243,7 @@ def custom_cat(id):
             new_cat_game = Category_Game(category_id=category_id, game_id=id)
             db.session.add(new_cat_game)
             db.session.commit()
+            flash('Game Added')
             return redirect ("/game-home")
     except:
         flash("Please select an option")
@@ -224,6 +251,7 @@ def custom_cat(id):
 
 @app.route('/category/<int:id>')
 def cat_list(id):
+    """Renders a custom category"""
     cat = Category.query.get(id)
     cat_game = Category_Game.query.filter_by(category_id=id).all()
     games = [c.game_id for c in cat_game]
@@ -232,6 +260,3 @@ def cat_list(id):
         game_data = Game.query.get_or_404(c)
         info.append(game_data)
     return render_template('cat_games.html', games=games, info=info,cat=cat)
-
-
-
